+++ app/config/color.js
@@ -1,0 +1,23 @@
+
+
+export default {
+    // Primary Colors
+    white: "#fff",
+
+    containerBg: "#E1E5EA",
+    balnceBg: "#64C9CF",
+    spendingsBg: "#588C73",
+    expenseListBg: "#F3F1F5",
+
+    // Button:
+    addBalanceBtn: "#4B6587",
+    addExpensesBtn: "#911F27",
+    btnTextColor: "#F3F0D7",
+    addMoneyBtn: "#1EAE98",
+    submitBtn: "#D83A56",
+    
+    amountFeildText: "#297F87",
+
+
+}
+