+++ app/nevigation/drawerContent.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import { Text } from 'react-native'
+
+
+
+function DrawerContent(props) {
+    return (
+        <Text>
+          Lorem ipsum dolor sit amet consectetur adipisicing elit. Tempore deserunt et laborum beatae quo ut reprehenderit omnis assumenda doloremque iste, natus corporis molestiae quidem sed quae, commodi quam cupiditate non nostrum repellat numquam. Praesentium labore quam non laborum, expedita, quia aliquid earum tenetur in, ipsam illo eligendi quidem sed. Laudantium quo possimus necessitatibus perferendis est magnam corrupti, exercitationem itaque molestiae debitis! Temporibus possimus corrupti qui nostrum repellendus, asperiores tempore corporis, distinctio laboriosam repellat necessitatibus ea nesciunt. Quaerat reprehenderit eum hic molestiae quibusdam, tempora incidunt deleniti eius, est iure amet suscipit?
+        </Text>
+    )
+}
+
+export default DrawerContent
+
+