+++ app/screens/allExpenses.js
@@ -4,7 +4,7 @@ import { StyleSheet, Text, View } from 'react-native'
 export default function AllExpenses() {
     return (
         <View>
-            <Text>Expenses</Text>
+            <Text>Lorem ipsum dolor sit amet consectetur, adipisicing elit. Reprehenderit nemo doloribus repellendus alias iusto cumque, a molestiae sit dolor rerum temporibus ducimus. Corporis laboriosam enim sit veritatis excepturi? Id, sed. Iusto magni nihil provident ipsa excepturi blanditiis dolorem consequuntur est architecto expedita ad odit obcaecati qui, officia ut quos error!</Text>
         </View>
     )
 }