+++ app/screens/allbalance.js
@@ -4,7 +4,7 @@ import { StyleSheet, Text, View } from 'react-native'
 export default function Allbalances() {
     return (
         <View>
-            <Text></Text>
+            <Text>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Illo reiciendis dolores architecto sapiente harum consectetur, eaque cum, repudiandae magnam deserunt exercitationem voluptas. Accusamus ipsum ratione ipsam libero facilis explicabo fuga commodi, officiis repellat reprehenderit debitis eveniet quia ipsa vero harum ea necessitatibus fugiat cumque autem? Ratione ducimus laudantium fugiat. Aliquid nam, repellendus voluptates facere ex reiciendis quisquam, qui quo nesciunt sequi dignissimos, natus illo illum numquam eius ullam velit omnis et dolor sint! Iste laboriosam totam nisi numquam quo sed, commodi unde corrupti possimus ipsum. Quod quibusdam tempora iusto incidunt ad magni unde vitae placeat, recusandae maxime, quia error totam.</Text>
         </View>
     )
 }